train_4

+++ train_with_positionembedding.py
@@ -388,8 +388,10 @@ def main(args):
             # --------------------------------------------------------------------------------------------------------- #
             # [4.0] classification loss
             # [4.1] total loss
-            normal_dist_max, normal_dist_loss = gen_mahal_loss(args, anormal_feat_list, normal_feat_list).to(weight_dtype)
-            down_dim_normal_dist_max, down_dim_normal_dist_loss = gen_mahal_loss(args, down_dim_anormal_feat_list, down_dim_normal_feat_list).to(weight_dtype)
+            normal_dist_max, normal_dist_loss = gen_mahal_loss(args, anormal_feat_list, normal_feat_list)
+            normal_dist_loss = normal_dist_loss.to(weight_dtype)
+            down_dim_normal_dist_max, down_dim_normal_dist_loss = gen_mahal_loss(args, down_dim_anormal_feat_list, down_dim_normal_feat_list)
+            down_dim_normal_dist_loss = down_dim_normal_dist_loss.to(weight_dtype)
 
             # logging -------------------------------------------------------------------------------------------------
